+++ airline-frontend/src/main/java/app/FrontendApplication.java
@@ -2,11 +2,14 @@ package app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-
+import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication
-public class PaymentsApplication {
+@EnableFeignClients
+public class FrontendApplication {
+
     public static void main(String[] args) {
-        SpringApplication.run(PaymentsApplication.class, args);
+        SpringApplication.run(FrontendApplication.class, args);
     }
+
 }