+++ airline-project/src/main/java/app/AirlineApplication.java
@@ -5,9 +5,9 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.web.bind.annotation.RestController;
 
-@SpringBootApplication
 @RestController
 @EnableFeignClients
+@SpringBootApplication
 public class AirlineApplication {
     public static void main(String[] args) {
         SpringApplication.run(AirlineApplication.class, args);