+++ common/src/main/java/app/controllers/api/rest/ExampleRestApi.java
@@ -1,37 +1,45 @@
 package app.controllers.api.rest;
 
 import app.dto.ExampleDto;
-import io.swagger.annotations.*;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import io.swagger.annotations.ApiParam;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestBody;
 
-import javax.validation.Valid;
+
 
 @Api(tags = "Example REST")
 @Tag(name = "Example REST", description = "API example description")
-@RequestMapping("/api/example")
 public interface ExampleRestApi {
 
-    @GetMapping
+    @RequestMapping(value = "/api/example/", method = RequestMethod.GET)
     @ApiOperation(value = "Get Page of Examples")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Examples Page found"),
             @ApiResponse(code = 204, message = "Examples Page not present")}
     )
-    ResponseEntity<Page<ExampleDto>> getPage(@ApiParam(name = "page") @RequestParam(required = false) Integer page,
-                                             @ApiParam(name = "size") @RequestParam(required = false) Integer size);
+    ResponseEntity<Page<ExampleDto>> getPage(@ApiParam(name = "page") @RequestParam(required = false, value = "page") Integer page,
+                                             @ApiParam(name = "size") @RequestParam(required = false, value = "size") Integer size);
 
-    @GetMapping("/{id}")
+    @RequestMapping(value = "/api/example/{id}", method = RequestMethod.GET)
     @ApiOperation(value = "Get Example by id")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Example found"),
             @ApiResponse(code = 404, message = "Example not found")}
     )
-    ResponseEntity<ExampleDto> get(@ApiParam(name = "id", value = "Example.id") @PathVariable Long id);
+    ResponseEntity<ExampleDto> get(@ApiParam(name = "id", value = "Example.id") @PathVariable(value = "id") Long id);
 
-    @PostMapping
+    @RequestMapping(value = "/api/example/", method = RequestMethod.POST)
     @ApiOperation(value = "Create Example")
     @ApiResponses(value = {
             @ApiResponse(code = 201, message = "Example created"),
@@ -39,20 +47,20 @@ public interface ExampleRestApi {
     )
     ResponseEntity<ExampleDto> create(@ApiParam(name = "example", value = "ExampleDto") @Valid @RequestBody ExampleDto exampleDto);
 
-    @PatchMapping("/{id}")
+    @RequestMapping(value = "/api/example/{id}", method = RequestMethod.PATCH)
     @ApiOperation(value = "Update Example")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Example updated"),
             @ApiResponse(code = 400, message = "Example not updated")}
     )
-    ResponseEntity<ExampleDto> update(@ApiParam(name = "id", value = "Example.id") @PathVariable Long id,
+    ResponseEntity<ExampleDto> update(@ApiParam(name = "id", value = "Example.id") @PathVariable(value = "id") Long id,
                                       @ApiParam(name = "example", value = "ExampleDto") @Valid @RequestBody ExampleDto exampleDto);
 
-    @DeleteMapping("/{id}")
+    @RequestMapping(value = "/api/example/{id}", method = RequestMethod.DELETE)
     @ApiOperation(value = "Delete Example by id")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Example deleted"),
             @ApiResponse(code = 404, message = "Example not found")}
     )
-    ResponseEntity<Void> delete(@ApiParam(name = "id", value = "Example.id") @PathVariable Long id);
+    ResponseEntity<Void> delete(@ApiParam(name = "id", value = "Example.id") @PathVariable(value = "id") Long id);
 }