+++ common/src/main/java/app/dto/ExampleDto.java
@@ -1,17 +1,18 @@
 package app.dto;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Size;
 import lombok.Data;
 import lombok.Setter;
+import org.springframework.data.annotation.ReadOnlyProperty;
 
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.Size;
+
 
 @Data
 @Setter
 public class ExampleDto {
 
-    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
+    @ReadOnlyProperty
     private Long id;
 
     @Size(min = 1, max = 255, message = "Length of Example's exampleText should be between 1 and 255 characters")