+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- #277 Обновлены версии зависимостей в pom файлах и исправлены конфликты.
 - #237 Исправлен баг с редактированием Ticket. Убраны аннотации @Data из проекта.
 - #257 Переход AccountMapper, BookingMapper на Mapstruct , а также junit тесты к ним.
 - #258 Добавлена автоматическая конфигурация OpenApi при запуске приложения.
@@ -11,6 +12,7 @@
 - #no_task - added vaadin
 - #278 Перенос мапперов из контроллера в сервис.
 - #285 Фронт для destination
+- #306 Добавлен новый микросервис с фронтом(Vaadin), ExampleDTO и ExampleApi перенесен в common. Весь фронт был перенесен на отдельный сервис в airline-frontend. 
 - #236 Исправлен баг UpdateFlightById, справлен FlightMapper (удален прежний, добавлен FlightMapper интерфейс с использованием MapStruct), написаны unit есты для FlightMapper, интеграционные тесты для FlightController (updateFlightById)
 - #294 Добавлен Mapper для Destination через Mapstruct
 ## [1.0.6] - 30.07.2023